ML/requirements.txt
# Machine Learning Dependencies
# Core ML libraries
pandas>=2.1.3
numpy>=1.25.2
scikit-learn>=1.3.2
xgboost>=2.0.2
joblib>=1.3.2

# Data visualization and analysis
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.27.0
notebook>=7.0.0

# Database connection
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23

# Feature engineering
feature-engine>=1.6.0

# Model evaluation and validation
scikit-plot>=0.3.7

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0

# Optional: Advanced ML
# lightgbm>=4.1.0
# catboost>=1.2.0
# optuna>=3.4.0  # Hyperparameter optimization

